README: add description how to separately build the toolchain
The buildroot internal toolchain can be easily re-used to build external
software like u-boot as well.  Add instructions to the README how the
toolchain can be built without pulling in the entire root filesystem and
the kernel.  These intructions can then be referenced in the Meerkat
User Manual (which, regarding software, is only concerned about the
bootloader and has nothing to do with the BSP otherwise).

Change-Id: I09f9125f830a6ad13f8b368d4b6c3884c6e5a09a
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/3499
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>