board: meerkat-l4t: Add script to flash rootfs
This script allows comfortable flashing of a diskimage onto the eMMC of a
Meerkat based system utilising the u-boot USB Mass Storage (ums) feature. For
now it creates a single partition on the device and disk-dumps the supplied
disk image onto it.
Later on we can expand it to also flash a bootloader, etc.

Change-Id: Ia325d9dfe8227a8b7b57cbed51e6634c94a85044
Signed-off-by: Julian Scheel <julian@jusst.de>
Reviewed-on: http://review.adnet.avionic-design.de/3595
Reviewed-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>