BSP: defconfig: switch kernel version to tagged release
Switch the kernel version from the git development branch to the latest
release tag.

It is probably wise not to use a branch here anyway, since it may create
the illusion for users that this will track the branch and automatically
pull updates, while in reality the checkout is done only once when
"downloading" the package, with no further updates.

Change-Id: Iec0c3dbcba78244e7e76cf7e3d1e4d9d6fab250c
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/3675
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>