WIP: meerkat-l4t-bsp_defconfig: prep for customer
Prepare pushing of a development snapshot for the customer by using the
kernel from github rather than local copy, pinning it to a fixed commit,
and disabling ccache.

tinycamera source currently is still local and not distributed;
customers need to get the sources manually, unpack to the filesystem,
and modify NVIDIA_CFVC_SITE in the buildroot makefile for package
nvidia-cfvc, package/l4t/nvidia-cfvc/nvidia-cfvc.mk, to point to
the unpacked sources.

Change-Id: I868f4774d2585230c980ff78adc8ac407881119f
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>