meerkat-l4t_defconfig: switch from rootfs tarball to ext4 image
Most users want to flash onto the target eMMC.  In that case, having a
filesystem image is nicer, especially since the flashing procedure can
then be automated with flash.sh.  The bsp defconfig already defaults to
a filesystem image, while the minimal config creates a tarball by
default.  Switch the minimal config to a filesytem as well.

Change-Id: I1436b3b1da0f1ea8982bf1cad57d2d56e3abb563
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/4002
Reviewed-by: Julian Scheel <julian@jusst.de>