nvidia-gst: Fix installation of omx config
The install command to create /etc/xdg in target was wrong and created the dir
in the staging root only. This combined with the fact that the install command
for the actual config files did not end with a slash let install decide to
copy the omx config to the file /etc/xdg when no other package created the
directory /etc/xdg before. Fix the actual directory creation and add trailing
slashes to make the intention of installing into the directory explicit.

Change-Id: I95af0b8b22107faea5d7094b2b1f7f1051289ced
Signed-off-by: Julian Scheel <julian@jusst.de>
Reviewed-on: http://review.adnet.avionic-design.de/4311
Reviewed-by: Bert van Hall <bert.vanhall@avionic-design.de>