BSP: e17: Disable screen blanking
Now that we have connman added to the BSP, which sets the system clock
via NTP, the screen blanking timeout in e17 is triggered by the clock
adjustment that happens shortly after the system has booted.  Fix that
by disabling screen blanking in e17 altogether.

Unfortunately, with our simple method of syncing ~default/.e17 from a
bootstrapping machine, this pulls in quite a lot of new and modified
files into the rootfs overlay.

Change-Id: I8f6d58fb9538ec203f0ecf82e7404b11786c246d
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/4488
Reviewed-by: Julian Scheel <julian@jusst.de>