rootfs: sshd: Explicitly enable root login with password
We want to allow root logins with password via ssh.  The sshd default
for PermitRootLogin used to be "yes", so no particular configuration was
required.  But that default has changed to "prohibit-password" as per
openssh v7.0, which was merged into buildroot 2015.08.

Fix that in the post-build script by un-commenting the PermitRootLogin
keyword in sshd_config and setting it to "yes" explicitly.  Doing this
in the global post-build script is slightly more obscure than using an
rootfs overlay file, but it has the advantage of preserving
modifications to sshd_config from buildroot upstream.

Change-Id: Icc1209cd6288cf7c59af260ae63a90d2498002c8
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/4510
Reviewed-by: Julian Scheel <julian@jusst.de>