BSP: rootfs: Disable compositing in e17
Compositing in e17 does not work, which results in an ugly error message
on startup.  Disable it in the default user's configuration.

This is again achieved by simply syncing ~default/.e from a
bootstrapping system where the default user has disabled compositing in
the GUI.  Unfortunately that method is rather opaque and clobbers lots
of binary e17 configuration files.

Change-Id: I06772c7348344b89ebe91bbbe909beeb68c965d8
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/4562
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>