rootfs: Add a script to allow mounting the modules from the NFS server
In our test setup we use an NFS root with various kernels that are
stored outside of the rootfs. To still allow loading kernel modules
add a script that parse the kernel command line for a /lib parameter.
If found and the share contains a directory named modules/$VERSION it
is mounted in /lib/modules/$VERSION.

Change-Id: Ie79feeec2ba8e382d5d7186ec1af64c06ca383da
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/4682
Reviewed-by: Bert van Hall <bert.vanhall@avionic-design.de>