updater: Add the defconfig and overlay for the Meerkat Updater
This build a basic uclibc system that run an update script.
This system is normaly run as initrd from an update medium like an
sd-card or an USB stick. The updater script will create a set of
partitions and unpack the provided tarballs, optionally a pre and post
install scripts can also be run.

Change-Id: Ibd5745d277c4a780bba3628b5ef7a3d08f16bcdf
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/5208
Reviewed-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>