README: Document how to re-pull updated kernels from github
Since the kernel versions are fixed to a branch name which only rarely
changes, only the first build will pull an up-to-date kernel from
github, while later builds in the same working directory will use a
cached copy of the kernel sources.  Document in the README how to force
pulling an up-to-date kernel in an existing build.

Change-Id: I24f5af2c0445bf124b26923e2806f6e43feed33a
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/5233
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>