rootfs: Add a script to use a per device /root and /home with NFS root
When using an NFS root the home directories are shared between all
machines using the same root. This has various undesired side effect,
in particular the shell histories get overwritten.

To avoid this add a script that create private home directories
depending on the network interface used for the NFS root. The script
first check if an NFS root is used and resolve the network interface
MAC address. It then bind mount the directories /opt/homes/$MAC/root
and /opt/homes/$MAC/home to /root and /home. If these directories
don't exists they are instancied by copying the original ones.

Change-Id: I451d5927a0824f4eb09faeac9e051d14c1ec4815
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/5490
Reviewed-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>