Add eeupdate32 package
This adds a package for the Intel eeupdate32 tool. Unfortunately this package
is not published to the public, nor do we have proper release tarballs. Thus
the source link points to the unpacked binary on Avionic Design internal
storage.

If this package is to be built by anyone outside of Avionic Design he would
have to provide the eeupdate32 binary to the dl-folder directly or amend this
packages Makefile.

Change-Id: I045a754511b196ac63e2e0e9a3d0e7b1ca16165e
Signed-off-by: Julian Scheel <julian@jusst.de>
Reviewed-on: http://review.adnet.avionic-design.de/5931
Reviewed-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>