cuda: Add cross-building toolchain for arm as host package
Contrary to the native cuda packages for arm by Nvidia, which are
provided as a single download that installs a local Debian package
repository, the cross packages by Nvidia are provided as separate
downloads.  Unfortunately, they duplicate the cuda packages for the
target.

Rather than grabbing all cross packages by Nvidia, this just downloads
and installs a minimal subset of the cross packages that is sufficient
to get a working toolchain when it's added to the existing, native cuda
buildroot package for arm.

Change-Id: I3ea8425c5b62a82719b5aef4d41157bd29b768e4
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/6392
Reviewed-by: Julian Scheel <julian@jusst.de>