patches/libevas: Remove explicit /usr/include reference
Add patches to fix the libevas package configure step. It included hardcoded
references to /usr/include. As the actual fix resides in a m4 file a
autoreconf is required to apply it. Unfortunately we cannot enable
autoreconfigure out of the buildroot external overlay, which is why a second
patch is added that actually contains the result of an autoreconf run.

Change-Id: I9c8b3a78779bc78d030368638189b0392de5d0ba
Signed-off-by: Julian Scheel <julian@jusst.de>
Reviewed-on: http://review.adnet.avionic-design.de/6438
Reviewed-by: Bert van Hall <bert.vanhall@avionic-design.de>