Record git info in kernelrelease string and os-release on the target
Add a linux package build hook that extracts the git commit info from
the git-export.commit file, if present, and puts it into .scmversion.
git-export.commit is expected to be a template file with git commit info
that's expanded by git-archive as per a export-subst attribute.

Since .scmversion is included into the kernelrelease string, it is
visible with uname(1).  This can preserve git commit info for kernel
sources that are pulled from a remote git, even though buildroot is
stripping the .git dir from the sources by git-archiving the target
commit.

For builds with local sources as per OVERRIDE_SRCDIR, do the same by
visiting the original source tree, which is hopefully a git repo, and
extracting the scmversion there.

Also get commit info for the buildroot-external.  Record both in
/etc/os-release on the target.

Change-Id: Ib744c21a81d24396ae411859392eb5ab0eef77d6
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/6455
Reviewed-by: Julian Scheel <julian@jusst.de>