cbootimage-configs: Save dfu helpers to u-boot environment
If the target emmc is partitioned by the u-boot flasher, we know it
has only a single partition.  Save the dfu_alt_info to the persistent
u-boot environment so that it can be re-used to download new images to
this partition with dfu.

Also add a convenience environment variable dfucmd_mmc0 that enters dfu
mode, exposing the emmc when executed with "run dfucmd_mmc0".

Change-Id: Ibf3c0d6e8af83db2b24e8917259d0c1e1f7c51fd
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/6528
Reviewed-by: Julian Scheel <julian@jusst.de>