cuda-samples: Move to L4T v21.5 source package
Move the CUDA samples to v21.5.

This no longer just copies the minimal set of prebuilt cuda-samples that
was once published as a rather unofficial tarball by NVIDIA, but does a
proper build of the full set of CUDA samples from source, as distributed
with L4T 21.5.

Download url extracted from JetPack v2.2.1.

Change-Id: I8ffc17a3b3666f427c3c88579b279236d06e8539
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/6696
Reviewed-by: Julian Scheel <julian@jusst.de>