opencv4tegra: pkgconfig: Link with cuda stub libraries if possible
If the cuda stub libraries are present, don't require the full libs.

A note on over-linking is in order.

The cuda dependencies are private to libopencv, which means that in the
.pc file, they should be in Requires.private.  But the cuda libs live in
non-standard paths, which, unless an application links to cuda by
itself with the correct search paths, is known to the linker only at
run-time, not at build-time.

The correct approach would be to emit -rpath-link directives so that
applications using libopencv need not link against the cuda libs.  But
for CUDA projects, that would require wrapping OPENCV_LIBS with
-Xlinker, which is a mess when automake is used.  Accepting some
over-linking is just way easier.

Change-Id: I0d8b1a16b40a88b1bb96d26c7535e468bdd4cae7
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/6699
Reviewed-by: Julian Scheel <julian@jusst.de>