tegrarcm: Remove package, now packaged by buildroot upstream
Upstream buildroot has added the tegrarcm package with commit e903e45,
which is part of release v2016.05.

Since the upstream package is building tegrarcm v1.7, while for the BSP
we are using a more recent git snaphshot, switching to the upstream
package requires overriding the tegrarcm version in the external.mk.

Change-Id: I1c24dd4252685d1de3ee8d14f1153afefa108c2e
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/7251
Reviewed-by: Julian Scheel <julian@jusst.de>