opencv4tegra: Fix dependency on libjpeg
In Config.in, the dependency on libjepg had some lowercase letters
sprinkled in, effectively disabling the selection of the libjpeg.

But the dependency on libjpeg wasn't correct anyways, since a package
should not directly depend on a particular jpeg implementation, which
libjpeg is. And in fact there's no need to do this, since in buildroot
libjpeg-turbo is configured to be abi-compatible to libjpeg8.

So, fix opencv4tegra to depend only on the virtual jpeg package.

Change-Id: Ic61d9a3379ac6d106d50d08b43b995ef99fe097b
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/6712
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>