meerkat-l4t_defconfig: Re-enable cuda and cuda-samples
Since for l4t 21.5 the CUDA samples are built from source, they now
depend on the CUDA cross toolchain, which wasn't enabled in the
stripped-down meerkat-l4t_defconfig. In turn, this disabled the cuda
package since it was only pulled in by cuda-samples by a
select-dependency.

Fix this by adding the CUDA toolchain.

Change-Id: Ia79bbb13aba6c0891efb0a6df6e6af082a4b03c1
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/6828
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>