README: Update kernel update instructions for l4t 21.5 kernel
The kernel for l4t 21.5 comes from a different git branch than that for
21.4, which makes for a different version string used in the generated
source tarball and build directory name.  Update the instructions
that describe how to pull a fresh kernel from github to cover this,
and switch to using pathname globs here so that we don't have to update
this on every change of the kernel version or branch.

Change-Id: I7a3e15b87ce76abb907fa76355d8e72b238ccb05
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/6897
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>