rootfs: Add support for using X with a mainline kernel
When using a mainline kernel the nouveau driver should be used instead
of the nvidia driver. To use the nvidia driver an xorg.conf is
required, with nouveau we can just run without any xorg.conf.

To support both we rename our current xorg.conf to xorg-l4t.conf and
add a script that dynamically create an xorg.conf symlink according to
the kernel driver used for the framebuffer device. The script only
runs if there is no xorg.conf or if it is a symlink, in this way it
won't overwrite any eventual custom xorg.conf created by the user.

Change-Id: I38c8e9f876410810925148792a02947603dc58cf
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/7230
Reviewed-by: Bert van Hall <bert.vanhall@avionic-design.de>