rootfs: Fix the temperature init script when no thermal zone exists
If the kernel doesn't provides the requiered thermal zone the script
would set the maximum temperature to 2°C, which power off the board
right away in most cases.

Fix the script to only setup the sensors for which there is a thermal
zone. While at it also ensure that no useless error message is
generated because some globs don't expand.

Change-Id: I44b941752c047b833a1f87738ac9515dc25e453a
Signed-off-by: Alban Bedel <alban.bedel@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/7229
Reviewed-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>