cuda-samples: Disable build of simpleMPI sample binary
We don't have an MPI compiler in buildroot, so the simpleMPI sample
can't be built. The Makefile for the sample gracefully skips the build
if it doesn't find an MPI compiler, which is good, but if there exist an
MPI compiler on the host, it will happily try to build with that, and
fail.  Avoid that by explicitly disabling the simpleMPI sample.

Change-Id: I2441ba2e017ed2b10586383b0846b6cc25a67f82
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>