libtbb: Fix choice of compiler for builds with old make
On Ubuntu 14.04 with make 3.81, building libtbb fails, since with that
make version, make variables specified on the commandline do not
override variable assignments in sub-makefiles that are called
recursively. Thus, specifying the compiler per variable assignment on
the commandline does not work.

Instead, patch the libtbb makefile to choose the compiler from the
standard make variables, CC and CXX.

Change-Id: I2fe2b01184e627c3b5b41d67560cce4141851ae0
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>