libtbb: Fix runtime detection for cross-build
The libtbb makefiles try to detect the "runtime", that is the versions
of the linux kernel, glibc version and gcc version used, but the
detection logic doesn't handle cross-builds.  Pass in the version info
as a commandline argument.  For this no detection logic is necessary,
since we can leverage existing buildroot variables.

Change-Id: I8b0275ea8e06dc3b3ee41d38adaac8e8a34bee72
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/7395
Reviewed-by: Julian Scheel <julian@jusst.de>