cbootimage-configs: Add patch fixing dependency makefile generation
The build dependencies of the emmc image are tracked by an automatically
created makefile snippet.  This makefile wasn't created by a dedicated
makefile rule but only when the rule building the emmc image was
invoked, which sometimes caused the build to fail.  Fix this by adding a
dedicated makefile rule to create the dependencies makefile.

Change-Id: I1bc4f773fe07dc38ccf4d08cba39a1ef549c9a44
Signed-off-by: Nikolaus Schulz <nikolaus.schulz@avionic-design.de>
Reviewed-on: http://review.adnet.avionic-design.de/7853
Reviewed-by: Alban Bedel <alban.bedel@avionic-design.de>